% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{format_helper}
\alias{format_helper}
\title{Helper Function for the Formatter}
\usage{
format_helper(list, time_label, tz)
}
\arguments{
\item{list}{A list or section of list from API response}

\item{time_label}{time column to be formatted to POSIXct}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A formatted tibble
}
\description{
A helper function
}
