% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_cancel_order}
\alias{ftx_cancel_order}
\title{FTX Cancel Order}
\usage{
ftx_cancel_order(key, secret, subaccount, order_id, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{order_id}{Numeric value of order ID}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, result is successful: "Order queued for cancellation"
}
\description{
FTX Cancel Order
}
