% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_cancel_order_clientid}
\alias{ftx_cancel_order_clientid}
\title{FTX Cancel Order by Client ID}
\usage{
ftx_cancel_order_clientid(key, secret, subaccount, client_id, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{client_id}{Numeric value of client order ID}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, result is successful: "Order queued for cancellation"
}
\description{
FTX Cancel Order by Client ID
}
