% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_coin_balances}
\alias{ftx_coin_balances}
\title{FTX Coin Balances}
\usage{
ftx_coin_balances(key, secret, accounts = c(), ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{accounts}{Optional parameter. A vector of client sub-accounts}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns balances coin balances of all accounts if accounts argument is left empty or returns the balances of the specified accounts.
}
\examples{
ftx_coin_balances(key, secret, accounts = c())
}
