% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_coin_markets}
\alias{ftx_coin_markets}
\title{FTX Coin Markets}
\usage{
ftx_coin_markets(key, secret, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}
}
\value{
A tibble
}
\description{
FTX Coin Markets
}
