% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_funding_payments}
\alias{ftx_funding_payments}
\title{FTX Funding Payments}
\usage{
ftx_funding_payments(key, secret, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}
}
\value{
A tibble
}
\description{
FTX Funding Payments
}
