% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_future_funding_rates}
\alias{ftx_future_funding_rates}
\title{FTX Future Funding Rates}
\usage{
ftx_future_funding_rates(
  key,
  secret,
  markets = c(),
  start_time = NA,
  end_time = NA,
  tz = "GMT",
  ...
)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{markets}{Vector of names of markets.}

\item{start_time}{POSIXct value from when to extract trades.}

\item{end_time}{POSIXct value up-to when to extract trades.}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, data: tibble
}
\description{
FTX Future Funding Rates
}
