% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_future_markets}
\alias{ftx_future_markets}
\title{FTX Future Markets}
\usage{
ftx_future_markets(key, secret, market = NA, tz = "GMT", ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{market}{Name of market}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns all types of futures on FTX
}
\examples{
ftx_future_markets(key, secret)
}
