% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_get_hourly_markets}
\alias{ftx_get_hourly_markets}
\title{FTX Get Hourly Markets}
\usage{
ftx_get_hourly_markets(
  markets,
  start_date = "2019-07-27",
  end_date = as.character(Sys.Date())
)
}
\arguments{
\item{markets}{name of the markets (vector of string)}

\item{start_date}{human readable start date in UTC (string)}

\item{end_date}{human readable end date in UTC (string)}
}
\value{
dataframe of prices if API call was successful
}
\description{
Get hourly resolution historical data of multiple FTX markets
}
\examples{
ftx_get_hourly_markets(c("BTC-PERP", "ETH-PERP"), 
                       start_date = "2019-07-20", end_date = "2022-06-30")

}
