% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_get_markets_history}
\alias{ftx_get_markets_history}
\title{FTX Get History All}
\usage{
ftx_get_markets_history(markets, resolution, start_date, end_date)
}
\arguments{
\item{markets}{name of the markets (vector of string)}

\item{resolution}{resolution of candles in seconds (60, 300, 900, 3600, 14400, 86400)}

\item{start_date}{human readable start date in UTC (string)}

\item{end_date}{human readable end date in UTC (string)}
}
\value{
dataframe of prices if API call was successful
}
\description{
Get historical data of multiple FTX markets
}
\examples{
ftx_get_markets_history(c("BTC-PERP", "ETH-PERP"), resolution = 300,
                       start_date = "2022-07-20", end_date = "2022-07-30")

}
