% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_get_multiple_markets}
\alias{ftx_get_multiple_markets}
\title{FTX Get Multiple Markets}
\usage{
ftx_get_multiple_markets(
  markets,
  resolution = 86400,
  start_time = "2019-07-20",
  end_time = "2022-06-30"
)
}
\arguments{
\item{markets}{name of the markets (vector of string)}

\item{resolution}{representing aggregation length in seconds (int)}

\item{start_time}{human readable start time in UTC (string)}

\item{end_time}{human readable end time in UTC (string)}
}
\value{
dataframe of prices if API call was successful
}
\description{
Get market data for multipole FTX markets.
}
\examples{
ftx_get_multiple_markets(c("BTC-PERP", "ETH-PERP"), 
                         resolution = 86400, 
                         start_time = "2019-07-20", 
                         end_time = "2022-06-30")
}
