% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_historical_prices}
\alias{ftx_historical_prices}
\title{FTX Historical Prices}
\usage{
ftx_historical_prices(
  key,
  secret,
  market,
  resolution,
  start_time,
  end_time,
  ...
)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{market}{Name of market}

\item{resolution}{Window length in seconds. options: 15, 60, 300, 900, 3600, 14400, 86400, or any multiple of 86400 up to 30*86400}

\item{start_time}{Optional parameter. Numeric value from when to extract prices.}

\item{end_time}{Optional parameter. Numeric value up-to when to extract prices.}
}
\value{
A tibble
}
\description{
FTX Historical Prices
}
