% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_init}
\alias{ftx_init}
\title{FTX Initialise Endpoint Value}
\usage{
ftx_init(rftx_base_url = NA)
}
\arguments{
\item{rftx_base_url}{Optional parameter. An endpoint value such as https://ftx.com or https://ftx.us}
}
\description{
Changes the default endpoint value. Default endpoint is https://ftx.com
}
