% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_markets}
\alias{ftx_markets}
\title{FTX Markets}
\usage{
ftx_markets()
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns information on all types of markets on FTX: spot, perpetual futures, expiring futures, and MOVE contracts. Examples for each type are BTC/USD, BTC-PERP, BTC-0626, and BTC-MOVE-1005.
}
\examples{
ftx_markets()
}
