% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_modify_order}
\alias{ftx_modify_order}
\title{FTX Modify Order}
\usage{
ftx_modify_order(
  key,
  secret,
  subaccount,
  order_id,
  size,
  price,
  tz = "GMT",
  ...
)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{order_id}{Numeric value of order ID}

\item{size}{Size of order}

\item{price}{Price of order}

\item{tz}{Timezone to display times in. Default is GMT.}

\item{...}{Additional parameters to pass to API request}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Modifies an order's size and price
}
\examples{
ftx_modify_order(key, secret, subaccount, order_id, size, price)
}
