% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_my_spot_borrow_history}
\alias{ftx_my_spot_borrow_history}
\title{FTX Spot Borrow History}
\usage{
ftx_my_spot_borrow_history(key, secret, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}
}
\value{
A tibble
}
\description{
FTX Spot Borrow History
}
