% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_open_orders}
\alias{ftx_open_orders}
\title{FTX Open Orders}
\usage{
ftx_open_orders(key, secret, subaccount, markets = c(), tz = "GMT", ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{markets}{Vector of names of markets.}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, data: tibble
}
\description{
FTX Open Orders
}
