% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_order_fills}
\alias{ftx_order_fills}
\title{FTX Orders Fills}
\usage{
ftx_order_fills(key, secret, market, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{market}{Name of market}
}
\value{
A tibble
}
\description{
FTX Orders Fills
}
