% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_order_status}
\alias{ftx_order_status}
\title{FTX Order Status}
\usage{
ftx_order_status(key, secret, order_id, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{order_id}{Numeric value of order ID}
}
\value{
A tibble
}
\description{
FTX Order Status
}
