% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_order_status_clientid}
\alias{ftx_order_status_clientid}
\title{FTX Order Status by Client ID}
\usage{
ftx_order_status_clientid(key, secret, subaccount, client_id, tz = "GMT", ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{client_id}{Numeric value of client order ID}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, data: tibble
}
\description{
FTX Order Status by Client ID
}
