% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_orderbook}
\alias{ftx_orderbook}
\title{FTX Orderbook}
\usage{
ftx_orderbook(key, secret, market = NA, depth = 5, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{market}{Name of market}

\item{depth}{Market depth. Min 1, Max 100, default 5}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, data: tibble
}
\description{
FTX Orderbook
}
