% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_orders_history}
\alias{ftx_orders_history}
\title{FTX Orders History}
\usage{
ftx_orders_history(key, secret, market, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{market}{Name of market}
}
\value{
A tibble
}
\description{
FTX Orders History
}
