% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_place_order}
\alias{ftx_place_order}
\title{FTX Place Order}
\usage{
ftx_place_order(
  key,
  secret,
  subaccount,
  market = NA,
  side = NA,
  price = NA,
  type = NA,
  size = NA,
  reduceOnly = FALSE,
  ioc = FALSE,
  postOnly = FALSE,
  clientId = NA,
  tz = "GMT",
  ...
)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{market}{Name of market}

\item{side}{"buy" or "sell"}

\item{price}{Numeric value. Send null for market orders.}

\item{type}{"limit" or "market"}

\item{size}{Size of order}

\item{reduceOnly}{optional; default is false}

\item{ioc}{optional; default is false}

\item{postOnly}{optional; default is false}

\item{clientId}{optional; client order id}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, data: tibble
}
\description{
FTX Place Order
}
