% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_positions}
\alias{ftx_positions}
\title{FTX Positions}
\usage{
ftx_positions(key, secret, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}
}
\value{
A tibble
}
\description{
FTX Positions
}
