% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_spot_lending_history}
\alias{ftx_spot_lending_history}
\title{FTX Spot Lending History}
\usage{
ftx_spot_lending_history(key, secret, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}
}
\value{
A tibble
}
\description{
FTX Spot Lending History
}
