% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_trades}
\alias{ftx_trades}
\title{FTX Orderbook}
\usage{
ftx_trades(key, secret, market, start_time, end_time, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{market}{Name of market}

\item{start_time}{Optional parameter. Numeric value from when to extract trades.}

\item{end_time}{Optional parameter. Numeric value up-to when to extract trades.}
}
\value{
A tibble
}
\description{
FTX Orderbook
}
