% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{result_formatter}
\alias{result_formatter}
\title{Result Formatter}
\usage{
result_formatter(result, time_label)
}
\arguments{
\item{result}{The result of an API request}

\item{time_label}{time column to be formatted to POSIXct}
}
\value{
A formatted tibble
}
\description{
Result Formatter
}
