% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_cancel_order_clientid}
\alias{ftx_cancel_order_clientid}
\title{FTX Cancel Order by Client ID}
\usage{
ftx_cancel_order_clientid(key, secret, subaccount, client_id, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{client_id}{Character string of client order ID}

\item{...}{Additional parameters to pass to API request}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
result: if success is TRUE: "Order queued for cancellation"
}
\description{
Queues an order for cancellation using the client ID instead of the order ID
}
\examples{
ftx_cancel_order_clientid(key="", secret="", subaccount=NA, client_id="order1")
}
