% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_coin_balances}
\alias{ftx_coin_balances}
\title{FTX Coin Balances}
\usage{
ftx_coin_balances(key, secret, accounts = c(), ...)
}
\arguments{
\item{key}{A client key}

\item{secret}{A client secret}

\item{accounts}{Optional parameter. A vector of client sub-accounts}
}
\value{
A tibble
}
\description{
FTX Coin Balances
}
