% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_coin_markets}
\alias{ftx_coin_markets}
\title{FTX Coin Markets}
\usage{
ftx_coin_markets(key, secret, tz = "GMT", ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: if success is FALSE, NA otherwise,
data: a tibble with the following fields: name, baseCurrency, quoteCurrency,
quoteVolume24h, change1h, change24h, changeBod, highLeverageFeeExempt, minProvideSize,
type, underlying, enabled, ask, bid, last, postOnly, price, priceIncrement, sizeIncrement,
restricted, volumeUsd24h
}
\description{
Returns information on all types of markets on FTX
}
\examples{
ftx_coin_markets(key, secret)
}
