% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_coin_markets}
\alias{ftx_coin_markets}
\title{FTX Coin Markets}
\usage{
ftx_coin_markets(key, secret, tz = "GMT", ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, data: tibble
}
\description{
FTX Coin Markets
}
