% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_coin_markets}
\alias{ftx_coin_markets}
\title{FTX Coin Markets}
\usage{
ftx_coin_markets(key, secret, tz = "GMT", ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{tz}{Timezone to display times in. Default is GMT.}

\item{...}{Additional parameters to pass to API request}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns information on all types of markets on FTX
}
\examples{
ftx_coin_markets(key, secret)
}
