% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_funding_payments}
\alias{ftx_funding_payments}
\title{FTX Funding Payments}
\usage{
ftx_funding_payments(
  key,
  secret,
  subaccount,
  start_time = NA,
  end_time = NA,
  tz = "GMT",
  ...
)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{start_time}{Optional parameter. POSIXct value from when to extract trades.}

\item{end_time}{Optional parameter. POSIXct value up-to when to extract trades.}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, data: tibble
}
\description{
FTX Funding Payments
}
