% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_future_funding_rates}
\alias{ftx_future_funding_rates}
\title{FTX Future Funding Rates}
\usage{
ftx_future_funding_rates(key, secret, market, start_time, end_time, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{market}{Name of market}

\item{start_time}{Numeric value from when to extract rates.}

\item{end_time}{Numeric value up-to when to extract rates.}
}
\value{
A tibble
}
\description{
FTX Future Funding Rates
}
