% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_future_markets}
\alias{ftx_future_markets}
\title{FTX Future Markets}
\usage{
ftx_future_markets(key, secret, market = NA, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{market}{Name of market}
}
\value{
A tibble
}
\description{
FTX Future Markets
}
