% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_historical_prices}
\alias{ftx_historical_prices}
\title{FTX Historical Prices}
\usage{
ftx_historical_prices(
  key,
  secret,
  market,
  resolution = 14400,
  start_time = NA,
  end_time = NA,
  tz = "GMT",
  ...
)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{market}{Name of market}

\item{resolution}{Window length in seconds. options: 15, 60, 300, 900, 3600, 14400, 86400, or any multiple of 86400 up to 30*86400}

\item{start_time}{Optional parameter. POSIXct value from when to extract trades.}

\item{end_time}{Optional parameter. POSIXct value up-to when to extract trades.}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns historical prices of expired futures
}
\examples{
ftx_historical_prices(key, secret, market = "1INCH/USD")
}
