% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_modify_order}
\alias{ftx_modify_order}
\title{FTX Modify Order}
\usage{
ftx_modify_order(key, secret, order_id, size, price, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{order_id}{Numeric value of order ID}

\item{size}{Size of order}

\item{price}{Price of order}
}
\value{
A tibble
}
\description{
FTX Modify Order
}
