% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_modify_order}
\alias{ftx_modify_order}
\title{FTX Modify Order}
\usage{
ftx_modify_order(
  key,
  secret,
  subaccount,
  order_id,
  size,
  price,
  tz = "GMT",
  ...
)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{order_id}{Numeric value of order ID}

\item{size}{Size of order}

\item{price}{Price of order}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, data: tibble
}
\description{
FTX Modify Order
}
