% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_modify_order_clientid}
\alias{ftx_modify_order_clientid}
\title{FTX Modify Order by Client ID}
\usage{
ftx_modify_order_clientid(
  key,
  secret,
  subaccount,
  client_id,
  new_client_id,
  size,
  price,
  ...,
  tz = "GMT"
)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{client_id}{Character string of client order ID}

\item{new_client_id}{Character string of new client order ID}

\item{size}{Size of order}

\item{price}{Price of order}

\item{...}{Additional parameters to pass to API request}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Modifies an order using the client ID instead of the order ID
}
\examples{
ftx_modify_order_clientid(key="", secret="", subaccount=NA, client_id="order1", 
new_client_id="order2", size=2, price=1)
}
