% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_my_spot_borrow_history}
\alias{ftx_my_spot_borrow_history}
\title{FTX Spot Borrow History}
\usage{
ftx_my_spot_borrow_history(
  key,
  secret,
  subaccount,
  start_time = NA,
  end_time = NA,
  tz = "GMT",
  ...
)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{start_time}{Optional parameter. POSIXct value from when to extract trades.}

\item{end_time}{Optional parameter. POSIXct value up-to when to extract trades.}

\item{tz}{Timezone to display times in. Default is GMT.}

\item{...}{Additional parameters to pass to API request}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns the coin borrow history for the user
}
\examples{
ftx_my_spot_borrow_history(key, secret, subaccount)
}
