% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_open_orders}
\alias{ftx_open_orders}
\title{FTX Open Orders}
\usage{
ftx_open_orders(key, secret, subaccount, markets = c(), tz = "GMT", ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{markets}{Vector of names of markets.}

\item{tz}{Timezone to display times in. Default is GMT.}

\item{...}{Additional parameters to pass to API request}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns information on the open orders for account or subaccount if specified
}
\examples{
ftx_open_orders(key, secret, subaccount, markets=c('CRV-PERP','XRP-PERP'))
}
