% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_open_orders}
\alias{ftx_open_orders}
\title{FTX Open Orders}
\usage{
ftx_open_orders(key, secret, market, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{market}{Name of market}
}
\value{
A tibble
}
\description{
FTX Open Orders
}
