% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_order_fills}
\alias{ftx_order_fills}
\title{FTX Orders Fills}
\usage{
ftx_order_fills(
  key,
  secret,
  subaccount,
  markets = c(),
  start_time = NA,
  end_time = NA,
  tz = "GMT",
  ...
)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{markets}{Vector of names of markets.}

\item{start_time}{Optional parameter. POSIXct value from when to extract trades.}

\item{end_time}{Optional parameter. POSIXct value up-to when to extract trades.}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns market fills
}
\examples{
ftx_order_fills(key, secret, subaccount, markets=c('CRV-PERP','XRP-PERP'))
}
