% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_order_status}
\alias{ftx_order_status}
\title{FTX Order Status}
\usage{
ftx_order_status(key, secret, subaccount, order_id, tz = "GMT", ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{order_id}{Numeric value of order ID}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, data: tibble
}
\description{
FTX Order Status
}
