% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_order_status}
\alias{ftx_order_status}
\title{FTX Order Status}
\usage{
ftx_order_status(key, secret, subaccount, order_id, ..., tz = "GMT")
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{order_id}{Numeric value of order ID}

\item{...}{Additional parameters to pass to API request}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns the status of orders
}
\examples{
ftx_order_status(key="", secret="", subaccount=NA, order_id=1234)
}
