% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_order_status_clientid}
\alias{ftx_order_status_clientid}
\title{FTX Order Status by Client ID}
\usage{
ftx_order_status_clientid(key, secret, subaccount, client_id, tz = "GMT", ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{client_id}{Numeric value of client order ID}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns the status of orders using the client ID instead of the order ID
}
\examples{
ftx_order_status_clientid(key, secret, subaccount, client_id)
}
