% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_orderbook}
\alias{ftx_orderbook}
\title{FTX Orderbook}
\usage{
ftx_orderbook(key, secret, market = NA, depth = 5, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{market}{Name of market}

\item{depth}{Market depth. Min 1, Max 100, default 5}

\item{...}{Additional parameters to pass to API request}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns the orderbook for the market specified
}
\examples{
ftx_orderbook(key="", secret="", market = "1INCH/USD", depth = 5)
}
