% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_orders_history}
\alias{ftx_orders_history}
\title{FTX Orders History}
\usage{
ftx_orders_history(key, secret, subaccount, ..., markets = c(), tz = "GMT")
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{...}{Additional parameters to pass to API request}

\item{markets}{Vector of names of markets.}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns the history of orders for the account or subaccount if specified
}
\examples{
ftx_orders_history(key="", secret="", subaccount=NA, markets=c('CRV-PERP','XRP-PERP'))
}
