% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_orders_history}
\alias{ftx_orders_history}
\title{FTX Orders History}
\usage{
ftx_orders_history(key, secret, subaccount, markets = c(), tz = "GMT", ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{markets}{Vector of names of markets.}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, data: tibble
}
\description{
FTX Orders History
}
