% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_place_order}
\alias{ftx_place_order}
\title{FTX Place Order}
\usage{
ftx_place_order(
  key,
  secret,
  subaccount,
  ...,
  market = NA,
  side = NA,
  price = NA,
  type = NA,
  size = NA,
  reduceOnly = FALSE,
  ioc = FALSE,
  postOnly = FALSE,
  client_id = NA,
  tz = "GMT"
)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{...}{Additional parameters to pass to API request}

\item{market}{Name of market}

\item{side}{"buy" or "sell"}

\item{price}{Numeric value. Send null for market orders.}

\item{type}{"limit" or "market"}

\item{size}{Size of order}

\item{reduceOnly}{optional; default is false}

\item{ioc}{optional; default is false}

\item{postOnly}{optional; default is false}

\item{client_id}{optional; client order id}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Places an order based on the information provided
}
\examples{
ftx_place_order(key="",secret="",subaccount=NA,market="XRP-PERP",
side="buy",price=1,type="limit",size=3)
}
