% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_positions}
\alias{ftx_positions}
\title{FTX Positions}
\usage{
ftx_positions(key, secret, subaccount, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, data: tibble
}
\description{
FTX Positions
}
