% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_positions}
\alias{ftx_positions}
\title{FTX Positions}
\usage{
ftx_positions(key, secret, subaccount = NA, tz = "GMT", ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: if success is FALSE, NA otherwise,
data: a tibble with the following fields: cost, cumulativeBuySize, cumulativeSellSize,
entryPrice, estimatedLiquidationPrice, future, initialMarginRequirement, longOrderSize,
maintenanceMarginRequirement, netSize, openSize, realizedPnl, recentAverageOpenPrice,
recentBreakEvenPrice, recentPnl, shortOrderSize, side, size, unrealizedPnl, collateralUsed
}
\description{
Returns the account positions
}
\examples{
ftx_positions(key, secret, subaccount=NA)
}
