% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_positions}
\alias{ftx_positions}
\title{FTX Positions}
\usage{
ftx_positions(key, secret, ..., subaccount = NA, tz = "GMT")
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{...}{Additional parameters to pass to API request}

\item{subaccount}{A client's subaccount}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A list of three elements: a logical vector success: FALSE/TRUE,
failure_reason: reason for failure if success is FALSE, NA otherwise,
data: a tibble containing the data if success is TRUE
}
\description{
Returns the account positions
}
\examples{
ftx_positions(key="", secret="", subaccount=NA)
}
