% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_send_request}
\alias{ftx_send_request}
\title{FTX Send Request}
\usage{
ftx_send_request(method, path, key, secret, subaccount, ...)
}
\arguments{
\item{method}{REST API Method such as GET or POST}

\item{path}{An additional path defined for each function}

\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}
}
\value{
A response returned by request using specified method.
}
\description{
FTX Send Request
}
