% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{ftx_trades}
\alias{ftx_trades}
\title{FTX Orderbook}
\usage{
ftx_trades(key, secret, subaccount, market, start_time, end_time, ...)
}
\arguments{
\item{key}{A client's key}

\item{secret}{A client's secret}

\item{subaccount}{A client's subaccount}

\item{market}{Name of market}

\item{start_time}{Optional parameter. POSIXct value from when to extract trades.}

\item{end_time}{Optional parameter. POSIXct value up-to when to extract trades.}
}
\value{
A list of three elements: success: false/true, failure_reason: if available, data: tibble
}
\description{
FTX Orderbook
}
