% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftx.R
\name{result_formatter}
\alias{result_formatter}
\title{Result Formatter}
\usage{
result_formatter(result, time_label, tz)
}
\arguments{
\item{result}{The result of an API response}

\item{time_label}{time column to be formatted to POSIXct}

\item{tz}{Timezone to display times in. Default is GMT.}
}
\value{
A formatted tibble
}
\description{
A helper function
}
